src/requirements.txt
alembic==1.13.1
amqp==5.2.0
annotated-types==0.6.0
anyio==4.3.0
asgi-lifespan==2.1.0
asgiref==3.7.2
asyncpg==0.29.0
billiard==4.2.0
black==24.3.0
celery==5.3.6
certifi==2024.2.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
databases==0.9.0
fastapi==0.110.0
filelock==3.13.1
fonttools==4.50.0
fsspec==2024.3.1
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
iniconfig==2.0.0
isort==5.13.2
Jinja2==3.1.3
kiwisolver==1.4.5
kombu==5.3.5
Mako==1.3.2
MarkupSafe==2.1.5
matplotlib==3.8.3
mpmath==1.3.0
mypy==1.9.0
mypy-extensions==1.0.0
networkx==3.2.1
numpy==1.26.4
packaging==24.0
pandas==2.2.1
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.2.0
pluggy==1.4.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
pydantic==2.6.4
pydantic_core==2.16.3
pyparsing==3.1.2
pytest==8.1.1
pytest-asyncio==0.23.6
pytest-celery==0.0.0
python-dateutil==2.9.0.post0
python-multipart==0.0.9
pytz==2024.1
redis==5.0.3
ruff==0.3.3
setuptools==69.2.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.28
sqlmodel==0.0.16
starlette==0.36.3
sympy==1.12
torch==2.2.1
torchsummary==1.5.1
torchvision==0.17.1
types-colorama==0.4.15.20240311
types-psycopg2==2.9.21.20240311
typing_extensions==4.10.0
tzdata==2024.1
uvicorn==0.28.1
vine==5.1.0
wcwidth==0.2.13
wheel==0.42.0
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.1
amqp==5.2.0
annotated-types==0.6.0
anyio==4.3.0
asgi-lifespan==2.1.0
asgiref==3.7.2
asyncer==0.0.5
asyncpg==0.29.0
attrs==23.2.0
bcrypt==4.1.2
billiard==4.2.0
black==24.3.0
celery==5.3.6
celery-sqlalchemy-scheduler==0.3.0
certifi==2024.2.2
cffi==1.16.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
contourpy==1.2.0
cryptography==42.0.5
cycler==0.12.1
databases==0.9.0
fastapi==0.110.0
fastapi-async-sqlalchemy==0.6.1
fastapi-cache2==0.2.1
fastapi-limiter==0.1.6
fastapi-pagination==0.12.22
filelock==3.13.1
fonttools==4.50.0
frozenlist==1.4.1
fsspec==2024.3.1
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
iniconfig==2.0.0
isort==5.13.2
Jinja2==3.1.3
kiwisolver==1.4.5
kombu==5.3.5
Mako==1.3.2
MarkupSafe==2.1.5
matplotlib==3.8.3
mpmath==1.3.0
multidict==6.0.5
mypy==1.9.0
mypy-extensions==1.0.0
networkx==3.2.1
numpy==1.26.4
packaging==24.0
pandas==2.2.1
pathspec==0.12.1
pendulum==3.0.0
pillow==10.2.0
platformdirs==4.2.0
pluggy==1.4.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
pycparser==2.22
pydantic==2.6.4
pydantic-settings==2.2.1
pydantic_core==2.16.3
pyparsing==3.1.2
pytest==8.1.1
pytest-asyncio==0.23.6
pytest-celery==0.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
redis==5.0.3
ruff==0.3.3
setuptools==69.2.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.28
SQLAlchemy-Utils==0.41.2
sqlmodel==0.0.16
starlette==0.36.3
sympy==1.12
time-machine==2.14.1
torch==2.2.1
torchsummary==1.5.1
torchvision==0.17.1
types-colorama==0.4.15.20240311
types-psycopg2==2.9.21.20240311
typing_extensions==4.10.0
tzdata==2024.1
uvicorn==0.28.1
vine==5.1.0
wcwidth==0.2.13
wheel==0.42.0
yarl==1.9.4
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.1
amqp==5.2.0
annotated-types==0.6.0
anyio==4.3.0
asgi-lifespan==2.1.0
asgiref==3.7.2
asyncer==0.0.5
asyncpg==0.29.0
attrs==23.2.0
bcrypt==4.1.2
billiard==4.2.0
black==24.3.0
celery==5.3.6
celery-sqlalchemy-scheduler==0.3.0
certifi==2024.2.2
cffi==1.16.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
contourpy==1.2.0
cryptography==42.0.5
cycler==0.12.1
databases==0.9.0
fastapi==0.110.0
fastapi-async-sqlalchemy==0.6.1
fastapi-cache2==0.2.1
fastapi-limiter==0.1.6
fastapi-pagination==0.12.22
filelock==3.13.1
fonttools==4.50.0
frozenlist==1.4.1
fsspec==2024.3.1
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
iniconfig==2.0.0
isort==5.13.2
Jinja2==3.1.3
kiwisolver==1.4.5
kombu==5.3.5
Mako==1.3.2
MarkupSafe==2.1.5
matplotlib==3.8.3
mpmath==1.3.0
multidict==6.0.5
mypy==1.9.0
mypy-extensions==1.0.0
networkx==3.2.1
numpy==1.26.4
packaging==24.0
pandas==2.2.1
pathspec==0.12.1
pendulum==3.0.0
pillow==10.2.0
platformdirs==4.2.0
pluggy==1.4.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
pycparser==2.22
pydantic==2.6.4
pydantic-settings==2.2.1
pydantic_core==2.16.3
PyJWT==2.8.0
pyparsing==3.1.2
pytest==8.1.1
pytest-asyncio==0.23.6
pytest-celery==0.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
redis==5.0.3
ruff==0.3.3
setuptools==69.2.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.28
SQLAlchemy-Utils==0.41.2
sqlmodel==0.0.16
starlette==0.36.3
sympy==1.12
time-machine==2.14.1
torch==2.2.1
torchsummary==1.5.1
torchvision==0.17.1
types-colorama==0.4.15.20240311
types-psycopg2==2.9.21.20240311
typing_extensions==4.10.0
tzdata==2024.1
uvicorn==0.28.1
vine==5.1.0
wcwidth==0.2.13
wheel==0.42.0
yarl==1.9.4
